requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
numpy
tensorflow-cpu
opencv-python-headless
requests
